Backport `#[]` and `#[]=` to Active Model
Migrate the `ActiveRecord::AttributeMethods#[]` [read][] method, the
`ActiveRecord::AttributeMethods#[]=` [write][] method, as well as the
underlying [ActiveRecord::AttributeMethods::Read#read_attribute][] and
[ActiveRecord::AttributeMethods::Write#write_attribute][] methods from
Active Record to Active Model.

This change preserves existing Active Record behavior, while introducing
a new read and write syntax to classes that mix-in
`ActiveModel::Attributes`. In addition to mimicking Active Record, there
are some other niceties like integration with [Enumerable#pluck][]:

```ruby
topics = [
  Topic.new(title: "First"),
  Topic.new(title: "Second")
]

topics.pluck(:title) # => ["First", "Second"]
```

Once in place, this change also enables [a potential back-port of
ActiveRecord::Normalization][ActiveRecord::Normalization] to Active
Model, since it [utilizes the braces syntax][write].

[read]: https://edgeapi.rubyonrails.org/classes/ActiveRecord/AttributeMethods.html#method-i-5B-5D
[write]: https://edgeapi.rubyonrails.org/classes/ActiveRecord/AttributeMethods.html#method-i-5B-5D-3D
[write_attribute]: https://edgeapi.rubyonrails.org/classes/ActiveRecord/AttributeMethods/Write.html#method-i-write_attribute
[read_attribute]: https://edgeapi.rubyonrails.org/classes/ActiveRecord/AttributeMethods/Read.html#method-i-read_attribute
[Enumerable#pluck]: https://edgeapi.rubyonrails.org/classes/Enumerable.html#method-i-pluck
[ActiveRecord::Normalization]: https://github.com/rails/rails/discussions/53793#discussioncomment-11500218
[write]: https://github.com/rails/rails/blob/main/activerecord/lib/active_record/normalization.rb#L28
