+++ activemodel/test/cases/attribute_methods_test.rb
@@ -52,6 +52,21 @@ protected
   end
 end
 
+class ModelWithAttributes3
+  include ActiveModel::AttributeMethods
+
+  attr_accessor :value
+
+  attribute_method_suffix "?"
+
+  define_attribute_methods :value
+
+  private
+    def attribute?(name)
+      self[name].present?
+    end
+end
+
 class ModelWithAttributesWithSpaces
   include ActiveModel::AttributeMethods
 
@@ -103,6 +118,14 @@ end
 
 class AttributeMethodsTest < ActiveModel::TestCase
   include ActiveSupport::Testing::RactorsAssertions
+
+  class Topic
+    include ActiveModel::AttributeMethods
+
+    attr_accessor :title
+
+    alias_attribute :heading, :title
+  end
 
   test "method missing works correctly even if attributes method is not defined" do
     assert_raises(NoMethodError) { ModelWithoutAttributesMethod.new.foo }
@@ -293,6 +316,16 @@ class AttributeMethodsTest < ActiveModel::TestCase
     assert_equal "bar", attrs["foo"]
   end
 
+  test "accessing a suffixed attribute through #[]" do
+    m = ModelWithAttributes3.new
+
+    assert_not_predicate m, :value?
+
+    m.value = true
+
+    assert_predicate m, :value?
+  end
+
   test "defined attribute doesn't expand positional hash argument" do
     ModelWithAttributes2.define_attribute_methods(:foo)
 
@@ -451,5 +484,100 @@ class AttributeMethodsTest < ActiveModel::TestCase
     end
 
     assert_ractor_shareable model.attribute_method_patterns
+  end
+
+  test "write_attribute" do
+    topic = Topic.new
+    topic.write_attribute :title, "Still another topic"
+    assert_equal "Still another topic", topic.title
+
+    topic[:title] = "Still another topic: part 2"
+    assert_equal "Still another topic: part 2", topic.title
+
+    topic.write_attribute "title", "Still another topic: part 3"
+    assert_equal "Still another topic: part 3", topic.title
+
+    topic["title"] = "Still another topic: part 4"
+    assert_equal "Still another topic: part 4", topic.title
+  end
+
+  test "write_attribute can write aliased attributes as well" do
+    topic = Topic.new
+    topic.title = "Don't change the topic"
+    topic.write_attribute :heading, "New topic"
+
+    assert_equal "New topic", topic.title
+  end
+
+  test "write_attribute raises ActiveModel::MissingAttributeError when the attribute does not exist" do
+    topic = Topic.new
+    assert_raises(ActiveModel::MissingAttributeError) { topic[:no_column_exists] = "Hello!" }
+  end
+
+  test "read_attribute" do
+    topic = Topic.new
+    topic.title = "Don't change the topic"
+    assert_equal "Don't change the topic", topic.read_attribute("title")
+    assert_equal "Don't change the topic", topic["title"]
+
+    assert_equal "Don't change the topic", topic.read_attribute(:title)
+    assert_equal "Don't change the topic", topic[:title]
+  end
+
+  test "read_attribute can read aliased attributes as well" do
+    topic = Topic.new
+    topic.title = "Don't change the topic"
+
+    assert_equal "Don't change the topic", topic.read_attribute("heading")
+    assert_equal "Don't change the topic", topic["heading"]
+
+    assert_equal "Don't change the topic", topic.read_attribute(:heading)
+    assert_equal "Don't change the topic", topic[:heading]
+  end
+
+  test "read_attribute raises ActiveModel::MissingAttributeError when the attribute does not exist" do
+    topic = Topic.new
+    assert_raises(ActiveModel::MissingAttributeError) { topic[:no_column_exists] }
+    assert_raises(ActiveModel::MissingAttributeError) { topic.read_attribute(:no_column_exists) }
+  end
+
+  test "overridden write_attribute" do
+    topic = Topic.new
+    def topic.write_attribute(attr_name, value)
+      super(attr_name, value.downcase)
+    end
+
+    topic.write_attribute :title, "Yet another topic"
+    assert_equal "yet another topic", topic.title
+
+    topic[:title] = "Yet another topic: part 2"
+    assert_equal "yet another topic: part 2", topic.title
+
+    topic.write_attribute "title", "Yet another topic: part 3"
+    assert_equal "yet another topic: part 3", topic.title
+
+    topic["title"] = "Yet another topic: part 4"
+    assert_equal "yet another topic: part 4", topic.title
+  end
+
+  test "overridden read_attribute" do
+    topic = Topic.new
+    topic.title = "Stop changing the topic"
+    def topic.read_attribute(attr_name)
+      super(attr_name).upcase
+    end
+
+    assert_equal "STOP CHANGING THE TOPIC", topic.read_attribute("title")
+    assert_equal "STOP CHANGING THE TOPIC", topic["title"]
+
+    assert_equal "STOP CHANGING THE TOPIC", topic.read_attribute(:title)
+    assert_equal "STOP CHANGING THE TOPIC", topic[:title]
+  end
+
+  test "non-attribute read and write" do
+    topic = Topic.new
+    assert_not_respond_to topic, "mumbo"
+    assert_raise(NoMethodError) { topic.mumbo }
+    assert_raise(NoMethodError) { topic.mumbo = 5 }
   end
 end
